Ignore nuget warnings for package shape
- error NU5103: The folder 'lib\
PackageManager\PackageManager.UI.exe' under 'lib' is not recognized as a valid framework name or a supported culture identifier. Rename itto a valid framework name or culture identifier.

- error NU5101: The assembly 'li
b\GitExtensions.PluginManager.dll' is placed directly under 'lib' folder. It is recommended that assemblies be placed inside a framework-specific folder. Move it into a framework-specific folder.

- error NU5128: Some target fram
eworks declared in the dependencies group of the nuspec and the lib/ref folder do not have exact matches in the other location. Consult the
 list of actions below: - Add a dependency group for PackageManager0.0 to the nuspec